Fix: user public data 와 user detail 분리
