+++ src/main/java/com/openbook/openbook/configuration/JwtAuthenticationFilter.java
@@ -3,22 +3,16 @@ package com.openbook.openbook.configuration;
 import com.openbook.openbook.global.exception.ErrorCode;
 import com.openbook.openbook.global.util.TokenProvider;
 import com.openbook.openbook.global.exception.OpenBookException;
-import com.openbook.openbook.user.dto.UserPublicData;
-import com.openbook.openbook.user.entity.User;
-import com.openbook.openbook.user.service.core.UserService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.constraints.NotNull;
 import java.io.IOException;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
