+++ src/main/java/com/openbook/openbook/user/dto/UserPublicData.java
@@ -1,64 +1,17 @@
 package com.openbook.openbook.user.dto;
 
 import com.openbook.openbook.user.entity.User;
-import java.util.Collection;
-import java.util.List;
-import net.minidev.json.annotate.JsonIgnore;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
 public record UserPublicData(
         Long id,
         String nickname,
         String role
-) implements UserDetails {
+) {
     public static UserPublicData of(User user) {
         return new UserPublicData(
                 user.getId(),
                 user.getNickname(),
                 user.getRole().name()
         );
-    }
-
-    @JsonIgnore
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority(role));
-    }
-
-    @JsonIgnore
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-    @Override
-    public String getUsername() {
-        return id.toString();
-    }
-
-    @JsonIgnore
-    @Override
-    public boolean isAccountNonExpired() {
-        return false;
-    }
-
-    @JsonIgnore
-    @Override
-    public boolean isAccountNonLocked() {
-        return false;
-    }
-
-    @JsonIgnore
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return false;
-    }
-
-    @JsonIgnore
-    @Override
-    public boolean isEnabled() {
-        return false;
     }
 }
